pyproject.toml
[project]
name = "strands-solver"
version = "0.1.0"
description = "A solver for Strands, the New York Times puzzle game."
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Nimo Beeren", email = "nimobeeren@gmail.com" }
]
license = { text = "MIT" }
keywords = ["strands", "nytimes", "puzzle-solver"]
dependencies = [
    "google-genai>=1.41.0",
    "httpx>=0.28.1",
    "mdpd>=0.2.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "sqlite-vec>=0.1.6",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "typer>=0.21.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
    "tiktoken>=0.11.0",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
strands-solver = "strands_solver.cli.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
markers = ["e2e: end-to-end tests that have external dependencies"]
addopts = "-m 'not e2e'"  # only run non-e2e tests by default
